Breaking up the pages to components
1. homepage
2. resume
